app_engine/demo/server/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
CacheControl==0.14.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
db-dtypes==1.2.0
dill==0.3.8
exceptiongroup==1.2.0
firebase-admin==6.4.0
Flask==3.0.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
flatbuffers==23.5.26
frozenlist==1.4.1
gast==0.5.4
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.42.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.21.0
google-cloud-bigquery-storage==2.24.0    # optional package, but without it, the slowdown of calling `.to_dataframe(...)` in the procedure `bq_client.query(sql).result().to_dataframe()` is 20x: https://github.com/googleapis/python-bigquery/issues/1492
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-resource-manager==1.12.1
google-cloud-secret-manager==2.23.0
google-cloud-storage==2.14.0
google-cloud-workflows==1.14.1
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.1
grpcio-status==1.60.1
gunicorn==21.2.0
h5py==3.10.0
holidays==0.42
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.15.0
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.16
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3    # pandas >= 2 has unpickling issues for dataframes that were constructed with pandas < 2
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydata-google-auth==1.8.2
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
pytest==8.0.0
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
shapely==2.0.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
threadpoolctl==3.3.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.14.1
yarl==1.9.4

## previously used packages
# google-cloud-appengine-logging==1.4.3    # no longer used because initialization of this package was causing request latency (see https://ficcai.atlassian.net/browse/FA-2022)
# google-cloud-logging==3.10.0    # no longer used because initialization of this package was causing request latency (see https://ficcai.atlassian.net/browse/FA-2022)
# pandarallel==1.6.5    # no longer used because the parallelization is handled by the `multiprocess` module and initialization of `pandarallel` was causing request latency (see https://ficcai.atlassian.net/browse/FA-2036)
# pickle5==0.0.11    # no longer used since we are currently using Python 3.10 and pickle5 was needed if Python <= 3.7
